+++ saleor/graphql/core/types/common.py
@@ -25,7 +25,13 @@ from ...core.doc_category import (
     DOC_CATEGORY_USERS,
     DOC_CATEGORY_WEBHOOKS,
 )
-from ..descriptions import ADDED_IN_36, ADDED_IN_312, DEPRECATED_IN_3X_FIELD
+from ..descriptions import (
+    ADDED_IN_36,
+    ADDED_IN_312,
+    ADDED_IN_313,
+    DEPRECATED_IN_3X_FIELD,
+    PREVIEW_FEATURE,
+)
 from ..enums import (
     AccountErrorCode,
     AppErrorCode,
@@ -459,7 +465,7 @@ class ProductVariantBulkError(Error):
     path = graphene.String(
         description=(
             "Path to field that caused the error. A value of `null` indicates that "
-            "the error isn't associated with a particular field."
+            "the error isn't associated with a particular field." + ADDED_IN_313
         ),
         required=False,
     )